rename: 일일 리포트 관련 클래스 폴더 변경
