+++ src/main/java/bsise/server/clovar/ClovaService.java
@@ -2,6 +2,7 @@ package bsise.server.clovar;
 
 import static bsise.server.user.domain.Preference.*;
 
+import bsise.server.clovar.dailyReport.ClovaDailyReportRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
