+++ src/main/java/bsise/server/clovar/dailyReport/AnalysisResult.java
@@ -1,4 +1,4 @@
-package bsise.server.clovar;
+package bsise.server.clovar.dailyReport;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
