+++ src/main/java/bsise/server/report/daily/DailyReportGetRequest.java
@@ -1,4 +1,4 @@
-package bsise.server.report;
+package bsise.server.report.daily;
 
 public interface DailyReportGetRequest {
 }
